+++ ipytest/_config.py
@@ -5,25 +5,27 @@ import warnings
 default_clean = "[Tt]est*"
 
 defaults = {
-    "rewrite_asserts": True,
-    "magics": True,
-    "clean": default_clean,
     "addopts": ("-q", "--color=yes"),
-    "run_in_thread": False,
+    "clean": default_clean,
+    "coverage": False,
     "defopts": "auto",
     "display_columns": 100,
+    "magics": True,
     "raise_on_error": False,
+    "rewrite_asserts": True,
+    "run_in_thread": False,
 }
 
 current_config = {
-    "rewrite_asserts": False,
-    "magics": False,
-    "clean": default_clean,
     "addopts": (),
-    "run_in_thread": False,
+    "clean": default_clean,
+    "coverage": False,
     "defopts": "auto",
     "display_columns": 100,
+    "magics": False,
     "raise_on_error": False,
+    "rewrite_asserts": False,
+    "run_in_thread": False,
 }
 
 _rewrite_transformer = None
@@ -67,6 +69,7 @@ def autoconfig(
     defopts=default,
     display_columns=default,
     raise_on_error=default,
+    coverage=default,
 ):
     """Configure `ipytest` with reasonable defaults.
 
@@ -91,6 +94,7 @@ def config(
     defopts=keep,
     display_columns=keep,
     raise_on_error=keep,
+    coverage=default,
 ):
     """Configure `ipytest`
 
@@ -105,6 +109,12 @@ def config(
     * `rewrite_asserts` (default: `False`): enable ipython AST transforms
       globally to rewrite asserts
     * `magics` (default: `False`): if set to `True` register the ipytest magics
+    * `coverage` (default: `False`): if `True` configure `pytest` to collect
+      coverage information. This functionality requires the `pytest-cov` package
+      to be installed. It adds `--cov --cov-config={GENERATED_CONFIG}` to the
+      arguments when invoking `pytest`. **WARNING**: this option will hide
+      existing coverage configuration files. See [`ipytest.cov`](#ipytestcov)
+      for details
     * `clean` (default: `[Tt]est*`): the pattern used to clean variables
     * `addopts` (default: `()`): pytest command line arguments to prepend to
       every pytest invocation. For example setting